Move rememeber the milk config storage to own module
