+++ homeassistant/components/remember_the_milk/const.py
@@ -1,0 +1,5 @@
+"""Constants for the Remember The Milk integration."""
+
+import logging
+
+LOGGER = logging.getLogger(__package__)
