+++ homeassistant/components/remember_the_milk/entity.py
@@ -1,14 +1,12 @@
 """Support to interact with Remember The Milk."""
 
-import logging
-
 from rtmapi import Rtm, RtmRequestFailedException
 
 from homeassistant.const import CONF_ID, CONF_NAME, STATE_OK
 from homeassistant.core import ServiceCall
 from homeassistant.helpers.entity import Entity
 
-_LOGGER = logging.getLogger(__name__)
+from .const import LOGGER
 
 
 class RememberTheMilkEntity(Entity):
@@ -24,7 +22,7 @@ class RememberTheMilkEntity(Entity):
         self._rtm_api = Rtm(api_key, shared_secret, "delete", token)
         self._token_valid = None
         self._check_token()
-        _LOGGER.debug("Instance created for account %s", self._name)
+        LOGGER.debug("Instance created for account %s", self._name)
 
     def _check_token(self):
         """Check if the API token is still valid.
@@ -34,7 +32,7 @@ class RememberTheMilkEntity(Entity):
         """
         valid = self._rtm_api.token_valid()
         if not valid:
-            _LOGGER.error(
+            LOGGER.error(
                 "Token for account %s is invalid. You need to register again!",
                 self.name,
             )
@@ -64,7 +62,7 @@ class RememberTheMilkEntity(Entity):
                 result = self._rtm_api.rtm.tasks.add(
                     timeline=timeline, name=task_name, parse="1"
                 )
-                _LOGGER.debug(
+                LOGGER.debug(
                     "Created new task '%s' in account %s", task_name, self.name
                 )
                 if hass_id is not None:
@@ -83,14 +81,14 @@ class RememberTheMilkEntity(Entity):
                     task_id=rtm_id[2],
                     timeline=timeline,
                 )
-                _LOGGER.debug(
+                LOGGER.debug(
                     "Updated task with id '%s' in account %s to name %s",
                     hass_id,
                     self.name,
                     task_name,
                 )
         except RtmRequestFailedException as rtm_exception:
-            _LOGGER.error(
+            LOGGER.error(
                 "Error creating new Remember The Milk task for account %s: %s",
                 self._name,
                 rtm_exception,
@@ -101,7 +99,7 @@ class RememberTheMilkEntity(Entity):
         hass_id = call.data[CONF_ID]
         rtm_id = self._rtm_config.get_rtm_id(self._name, hass_id)
         if rtm_id is None:
-            _LOGGER.error(
+            LOGGER.error(
                 (
                     "Could not find task with ID %s in account %s. "
                     "So task could not be closed"
@@ -120,11 +118,9 @@ class RememberTheMilkEntity(Entity):
                 timeline=timeline,
             )
             self._rtm_config.delete_rtm_id(self._name, hass_id)
-            _LOGGER.debug(
-                "Completed task with id %s in account %s", hass_id, self._name
-            )
+            LOGGER.debug("Completed task with id %s in account %s", hass_id, self._name)
         except RtmRequestFailedException as rtm_exception:
-            _LOGGER.error(
+            LOGGER.error(
                 "Error creating new Remember The Milk task for account %s: %s",
                 self._name,
                 rtm_exception,
